+++ src/utils/serviceAgreementMessage.ts
@@ -1,5 +1,28 @@
 import { formatUnixMsAsDate } from "@/utils/formatters"
 
+/// timeSigned is embedded in the wallet-signed message, so the server cannot
+/// replace it with its own clock - it can only refuse claims that disagree
+/// with it by more than the signing-ceremony window. The window is wide
+/// because Safe threshold signatures are legitimately proposed days before
+/// they are submitted; it still prevents arbitrary backdating of the stored
+/// legal record and the far-future timeSigned that would permanently win the
+/// monotonic replace guard in saveServiceAgreementSignature/Refusal.
+/// Lives here (not lib/) so client flows can align pending-signature expiry
+/// with the server's acceptance window.
+export const SERVICE_AGREEMENT_TIME_SIGNED_MAX_AGE_MS = 14 * 24 * 60 * 60 * 1000
+export const SERVICE_AGREEMENT_TIME_SIGNED_MAX_FUTURE_MS = 5 * 60 * 1000
+
+export function isServiceAgreementTimeSignedInBounds(
+  timeSigned: number,
+  now: number = Date.now(),
+): boolean {
+  return (
+    Number.isFinite(timeSigned) &&
+    timeSigned >= now - SERVICE_AGREEMENT_TIME_SIGNED_MAX_AGE_MS &&
+    timeSigned <= now + SERVICE_AGREEMENT_TIME_SIGNED_MAX_FUTURE_MS
+  )
+}
+
 export const normalizeServiceAgreementDeclineReason = (reason?: string) =>
   reason?.trim() || undefined
